Create Portfolio_Project_1
Google Data Analytics Capstone Project.
Synopsis: The project is about a fictional bike-sharing company, Cyclistic. The company provides 2 membership types: annual memberships as well as one day passes i.e. casual memberships. I was tasked with analyzing the company's trip data from the past 12 months to provide some insights on how the service is used differently among the two consumer groups. The overarching business goal is to develop an effective marketing campaign that appeals to the casual members to become annual members. With this in mind, I decided to do the entire cleaning and analyzing portion of the project on SQL.